+++ src/main/java/katapi/infrastructure/product/service/ProductService.java
@@ -1,14 +1,14 @@
 package katapi.infrastructure.product.service;
 
 import katapi.domain.product.Product;
-import katapi.infrastructure.product.exception.ProductNotFoundException;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductService {
 
     public List<Product> getAllProducts();
     public Product getProductById(Long id);
-    public Product createProduct(String name, Double price, Double weight);
+    public Product createProduct(String name, BigDecimal price, BigDecimal weight);
     public void deleteProductFromItsID(Long id);
 }
